Roll BuildProgress back into BuildEyebrow and update stories
